print_client/requirements.txt
websockets==13.0
python-escpos==3.1
